Created bot foler and moved all the .py files, besides main, for easier navigation in the project folder.
