+++ bot/message_syntax_functions.py
